mcp_gdrive/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.3
google-api-python-client==2.154.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
httpx==0.27.2
